# swagger_v1.yaml
openapi: 3.0.3
info:
  title: TravelGuide API
  description: API для управления городами и пользователями
  version: 2.0.0
servers:
  - url: http://localhost:8800/api/v2
    description: Development server
paths:
...
  /api/v1/cities:
    post:
      summary: Создать новый город
      description: Создает новый город в системе
      operationId: create_city
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Название города
                  example: "Москва"
              required:
                - name
      responses:
        '201':
          description: Город успешно создан
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - cities

    get:
      summary: Получить все города
      description: Возвращает список всех городов
      operationId: get_all_cities
      responses:
        '201':
          description: Список городов получен успешно
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - cities
  /api/v1/cities/{city_id}:
    get:
      summary: Получить город по ID
      description: Возвращает информацию о конкретном городе
      operationId: get_city
      parameters:
        - name: city_id
          in: path
          required: true
          description: ID города
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '201':
          description: Информация о городе получена успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityResponse'
        '404':
          description: Город не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - cities

    put:
      summary: Обновить город
      description: Обновляет информацию о городе
      operationId: update_city
      parameters:
        - name: city_id
          in: path
          required: true
          description: ID города
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Новое название города
                  example: "Санкт-Петербург"
              required:
                - name
      responses:
        '201':
          description: Город успешно обновлен
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Город не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - cities

    delete:
      summary: Удалить город
      description: Удаляет город по ID
      operationId: delete_city
      parameters:
        - name: city_id
          in: path
          required: true
          description: ID города
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Город успешно удален
        '404':
          description: Город не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - cities

  # ==================== ПОЛЬЗОВАТЕЛИ ====================
...
  /api/v1/login:
    post:
      summary: Войти в систему
      description: Аутентификация пользователя в системе
      operationId: login_user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                login:
                  type: string
                  description: Логин пользователя
                  example: "user123"
                password:
                  type: string
                  format: password
                  description: Пароль пользователя
                  example: "mypassword123"
              required:
                - login
                - password
      responses:
        '201':
          description: Успешный вход в систему
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT токен доступа
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user_id:
                    type: integer
                    description: ID пользователя
                    example: 1
                  message:
                    type: string
                    description: Сообщение об успехе
                    example: "Вход выполнен успешно"
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - authentication

  # ==================== ПРОФИЛИ ====================
...
  /api/v1/accommodations:
    get:
      summary: Get all accommodations
      description: Retrieve all accommodations and render HTML page
      operationId: get_all_accommodations
      tags:
        - accommodations
      responses:
        '201':
          description: HTML page with accommodations
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new accommodation
      description: Create a new accommodation and return HTML response
      operationId: create_accommodation
      tags:
        - accommodations
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                check_in:
                  type: string
                  format: date-time
                check_out:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Accommodation created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
  /api/v1/accommodations/{accommodation_id}:
    get:
      summary: Get accommodation details
      description: Retrieve detailed information about a specific accommodation
      operationId: get_accommodation
      tags:
        - accommodations
      parameters:
        - name: accommodation_id
          in: path
          required: true
          description: ID of the accommodation
          schema:
            type: integer
            minimum: 1
      responses:
        '201':
          description: Successful response
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Accommodation not found
        '500':
          description: Internal server error

    put:
      summary: Update accommodation
      description: Update an existing accommodation and return HTML response
      operationId: update_accommodation
      tags:
        - accommodations
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                check_in:
                  type: string
                  format: date-time
                check_out:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Accommodation updated successfully
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Accommodation not found
        '500':
          description: Internal server error

    patch:
      summary: Partial update accommodation dates
      description: Update only the check-in and check-out dates of an accommodation
      operationId: update_accommodation_dates
      tags:
        - accommodations
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                check_in:
                  type: string
                  format: date-time
                check_out:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Accommodation dates updated successfully
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Accommodation not found
        '500':
          description: Internal server error

    delete:
      summary: Delete accommodation
      description: Delete an accommodation and redirect to accommodations page
      operationId: delete_accommodation
      tags:
        - accommodations
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Accommodation not found
        '500':
          description: Internal server error
  /api/v1/route/{route_id}/accommodations/{accommodation_id}:
    delete:
      summary: Delete accommodation from route
      description: Delete an accommodation from a specific route and redirect to route edit page
      operationId: delete_accommodation_for_route
      tags:
        - route-accommodations
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: route_id
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Accommodation or route not found
        '500':
          description: Internal server error
    post:
      summary: Add accommodation to route
      description: Create a new accommodation and add it to a specific route
      operationId: add_acc_to_route
      tags:
        - route-accommodations
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                check_in:
                  type: string
                  format: date-time
                check_out:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '400':
          description: Bad request
        '404':
          description: Route not found
        '500':
          description: Internal server error
# ==================== РАЗВЛЕЧЕНИЕ ====================
...